Squashed commits: OpenSSL updates and comprehensive improvements
Combined all commits from d3ad7ff9ad53124db5becebb588440e32250be60 to upstream/master:
- Updated OpenSSL functions for compatibility with newer versions
- Fixed compiler warnings across multiple GCC versions
- Improved CI/CD pipeline (Travis CI, Cirrus CI)
- Enhanced TLS/SSL functionality and testing infrastructure
- Added support for TLSv1.3 and OpenSSL 3.x compatibility
- Thread-safe implementations (Slab, StatList with spinlock support)
- Added Context Support to Thread-Safe Stat List Iterators
- Pthread Thread Local Storage support
- Code modernization, cleanup, and uncrustify formatting
- Build system improvements and bug fixes
- Enhanced error handling and logging
- Performance optimizations and memory leak fixes
- SSL boundaries fixes and config improvements

Note: SSL test certificates excluded from this repository

+++ usual/pgutil.c
@@ -278,7 +278,7 @@ failed:
 #define long uintptr_t
 
 /* include gperf code */
-const char *pg_keyword_lookup_real(const char *str, unsigned int len);
+const char *pg_keyword_lookup_real(const char *str, size_t len);
 #include <usual/pgutil_kwlookup.h>
 
 bool pg_is_reserved_word(const char *str)
@@ -286,4 +286,3 @@ bool pg_is_reserved_word(const char *str)
 	const char *kw = pg_keyword_lookup_real(str, strlen(str));
 	return kw != NULL;
 }
-
